+++ packages/playground/src/createConfigDropdown.ts
@@ -1,5 +1,5 @@
-type Sandbox = import('typescript-sandbox').Sandbox
-type Monaco = typeof import('monaco-editor')
+type Sandbox = import("typescript-sandbox").Sandbox
+type Monaco = typeof import("monaco-editor")
 
 type OptionsSummary = {
   display: string
@@ -9,16 +9,17 @@ type OptionsSummary = {
   categoryDisplay: string
 }
 
+// This is where all the localized descriptions come from
 declare const optionsSummary: OptionsSummary[]
 
 export const createConfigDropdown = (sandbox: Sandbox, monaco: Monaco) => {
-  const configContainer = document.getElementById('config-container')!
-  const container = document.createElement('div')
-  container.id = 'boolean-options-container'
+  const configContainer = document.getElementById("config-container")!
+  const container = document.createElement("div")
+  container.id = "boolean-options-container"
   configContainer.appendChild(container)
 
   const compilerOpts = sandbox.getCompilerOptions()
-  const boolOptions = Object.keys(compilerOpts).filter(k => typeof compilerOpts[k] === 'boolean')
+  const boolOptions = Object.keys(compilerOpts).filter(k => typeof compilerOpts[k] === "boolean")
 
   // we want to make sections of categories
   const categoryMap = {} as { [category: string]: { [optID: string]: OptionsSummary } }
@@ -32,18 +33,18 @@ export const createConfigDropdown = (sandbox: Sandbox, monaco: Monaco) => {
   })
 
   Object.keys(categoryMap).forEach(categoryID => {
-    const categoryDiv = document.createElement('div')
-    const header = document.createElement('h4')
-    const ol = document.createElement('ol')
+    const categoryDiv = document.createElement("div")
+    const header = document.createElement("h4")
+    const ol = document.createElement("ol")
 
     Object.keys(categoryMap[categoryID]).forEach(optID => {
       const optSummary = categoryMap[categoryID][optID]
       header.textContent = optSummary.categoryDisplay
 
-      const li = document.createElement('li')
-      const label = document.createElement('label')
-      label.style.position = 'relative'
-      label.style.width = '100%'
+      const li = document.createElement("li")
+      const label = document.createElement("label")
+      label.style.position = "relative"
+      label.style.width = "100%"
 
       const svg = `<?xml version="1.0" encoding="UTF-8"?><svg width="20px" height="20px" viewBox="0 0 20 20" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
           <g stroke="none" stroke-width="1" fill="none" fill-rule="evenodd"
@@ -53,11 +54,11 @@ export const createConfigDropdown = (sandbox: Sandbox, monaco: Monaco) => {
       </svg>`
       label.innerHTML = `<span>${optSummary.id}</span><a href='../tsconfig#${optSummary.id}' class='compiler_info_link' alt='Look up ${optSummary.id} in the TSConfig Reference'>${svg}</a><br/>${optSummary.oneliner}`
 
-      const input = document.createElement('input')
+      const input = document.createElement("input")
       input.value = optSummary.id
-      input.type = 'checkbox'
+      input.type = "checkbox"
       input.name = optSummary.id
-      input.id = 'option-' + optSummary.id
+      input.id = "option-" + optSummary.id
 
       input.onchange = () => {
         sandbox.updateCompilerSetting(optSummary.id, input.checked)
@@ -75,26 +76,26 @@ export const createConfigDropdown = (sandbox: Sandbox, monaco: Monaco) => {
     container.appendChild(categoryDiv)
   })
 
-  const dropdownContainer = document.getElementById('compiler-dropdowns')!
+  const dropdownContainer = document.getElementById("compiler-dropdowns")!
 
-  const target = optionsSummary.find(sum => sum.id === 'target')!
+  const target = optionsSummary.find(sum => sum.id === "target")!
   const targetSwitch = createSelect(
     target.display,
-    'target',
+    "target",
     target.oneliner,
     sandbox,
     monaco.languages.typescript.ScriptTarget
   )
   dropdownContainer.appendChild(targetSwitch)
 
-  const jsx = optionsSummary.find(sum => sum.id === 'jsx')!
-  const jsxSwitch = createSelect(jsx.display, 'jsx', jsx.oneliner, sandbox, monaco.languages.typescript.JsxEmit)
+  const jsx = optionsSummary.find(sum => sum.id === "jsx")!
+  const jsxSwitch = createSelect(jsx.display, "jsx", jsx.oneliner, sandbox, monaco.languages.typescript.JsxEmit)
   dropdownContainer.appendChild(jsxSwitch)
 
-  const modSum = optionsSummary.find(sum => sum.id === 'module')!
+  const modSum = optionsSummary.find(sum => sum.id === "module")!
   const moduleSwitch = createSelect(
     modSum.display,
-    'module',
+    "module",
     modSum.oneliner,
     sandbox,
     monaco.languages.typescript.ModuleKind
@@ -104,10 +105,10 @@ export const createConfigDropdown = (sandbox: Sandbox, monaco: Monaco) => {
 
 export const updateConfigDropdownForCompilerOptions = (sandbox: Sandbox, monaco: Monaco) => {
   const compilerOpts = sandbox.getCompilerOptions()
-  const boolOptions = Object.keys(compilerOpts).filter(k => typeof compilerOpts[k] === 'boolean')
+  const boolOptions = Object.keys(compilerOpts).filter(k => typeof compilerOpts[k] === "boolean")
 
   boolOptions.forEach(opt => {
-    const inputID = 'option-' + opt
+    const inputID = "option-" + opt
     const input = document.getElementById(inputID) as HTMLInputElement
     input.checked = !!compilerOpts[opt]
   })
@@ -115,11 +116,11 @@ export const updateConfigDropdownForCompilerOptions = (sandbox: Sandbox, monaco:
   const compilerIDToMaps: any = {
     module: monaco.languages.typescript.ModuleKind,
     jsx: monaco.languages.typescript.JsxEmit,
-    target: monaco.languages.typescript.ScriptTarget,
+    target: monaco.languages.typescript.ScriptTarget
   }
 
   Object.keys(compilerIDToMaps).forEach(flagID => {
-    const input = document.getElementById('compiler-select-' + flagID) as HTMLInputElement
+    const input = document.getElementById("compiler-select-" + flagID) as HTMLInputElement
     const currentValue = compilerOpts[flagID]
     const map = compilerIDToMaps[flagID]
     // @ts-ignore
@@ -132,13 +133,13 @@ export const updateConfigDropdownForCompilerOptions = (sandbox: Sandbox, monaco:
 }
 
 const createSelect = (title: string, id: string, blurb: string, sandbox: Sandbox, option: any) => {
-  const label = document.createElement('label')
-  const textToDescribe = document.createElement('span')
-  textToDescribe.textContent = title + ':'
+  const label = document.createElement("label")
+  const textToDescribe = document.createElement("span")
+  textToDescribe.textContent = title + ":"
   label.appendChild(textToDescribe)
 
-  const select = document.createElement('select')
-  select.id = 'compiler-select-' + id
+  const select = document.createElement("select")
+  select.id = "compiler-select-" + id
   label.appendChild(select)
 
   select.onchange = () => {
@@ -151,18 +152,18 @@ const createSelect = (title: string, id: string, blurb: string, sandbox: Sandbox
     .filter(key => isNaN(Number(key)))
     .forEach(key => {
       // hide Latest
-      if (key === 'Latest') return
+      if (key === "Latest") return
 
-      const option = document.createElement('option')
+      const option = document.createElement("option")
       option.value = key
       option.text = key
 
       select.appendChild(option)
     })
 
-  const span = document.createElement('span')
+  const span = document.createElement("span")
   span.textContent = blurb
-  span.classList.add('compiler-flag-blurb')
+  span.classList.add("compiler-flag-blurb")
   label.appendChild(span)
 
   return label